+++ Samples/Omni/Source/DynamicSprites.cpp
@@ -55,8 +55,8 @@ std::map<wchar_t, symbol_t> palette =
 
 void TestDynamicSprites()
 {
-	terminal_set("window.title='Omni: dynamic sprites'");
-	terminal_set("U+E000: ../Media/Tiles.png, size=32x32, align=top-left");
+    TerminalSet( "window.title='Omni: dynamic sprites'" );
+    TerminalSet( "U+E000: ../Media/Tiles.png, size=32x32, align=top-left" );
 
 	int map_width = map[0].length();
 	int map_height = map.size();
@@ -164,5 +164,5 @@ void TestDynamicSprites()
 		}
 	}
 
-	terminal_set("U+E000: none; U+E100: none;");
+    TerminalSet( "U+E000: none; U+E100: none;" );
 }