+++ Samples/Omni/Source/Mouse.cpp
@@ -16,8 +16,8 @@ double_click_area = {1, 11, 17, 4}; // FIXME: common rectangle struct
 
 void TestMouse()
 {
-	terminal_set("window.title='Omni: mouse input'");
-	terminal_set("input.filter={keyboard, mouse+}");
+    TerminalSet( "window.title='Omni: mouse input'" );
+    TerminalSet( "input.filter={keyboard, mouse+}" );
     TerminalComposition( TK_ON );
 
 	bool precise_mouse = false;
@@ -178,5 +178,5 @@ void TestMouse()
 
 	terminal_color("white");
     TerminalComposition( TK_OFF );
-	terminal_set("input: precise-mouse=false, mouse-cursor=true, filter={keyboard}");
+    TerminalSet( "input: precise-mouse=false, mouse-cursor=true, filter={keyboard}" );
 }